AOP 폴더 및 LoggingAspect 클래스 생성 - 컨트롤러 및 서비스 메소드 실행 전/후

+++ src/main/java/com/ieumsae/chat/repository/GroupChatEntranceLogRepository.java
@@ -11,11 +11,11 @@ import java.util.Optional;
 public interface GroupChatEntranceLogRepository extends JpaRepository<GroupChatEntranceLog, Long> {
 
     // 이전에 접속 기록이 있는지 확인하는 메소드
-    boolean existsByChatIdxAndUserIdx(Integer chatIdx, Integer userIdx);
+    boolean existsByChatIdxAndUserIdx(Long chatIdx, Long userIdx);
 
     // 특정 그룹 채팅방에 특정 사용자가 가장 최근에 입장한 기록을 조회
-    Optional<GroupChatEntranceLog> findFirstByChatIdxAndUserIdxOrderByEntranceDateTimeDesc(Integer groupChatIdx, Integer userIdx);
+    Optional<GroupChatEntranceLog> findFirstByChatIdxAndUserIdxOrderByEntranceDateTimeDesc(Long groupChatIdx, Long userIdx);
 
     // 로그 테이블에 저장 전, 이미 기록이 있는지 확인하는 메소드
-    Optional<GroupChatEntranceLog> findByChatIdxAndUserIdx(Integer chatIdx, Integer userIdx);
+    Optional<GroupChatEntranceLog> findByChatIdxAndUserIdx(Long chatIdx, Long userIdx);
 }
